// perf/efps/tests/synthetic/sanity.types.ts
export type Slug = {
  _type: 'slug'
  current?: string
  source?: string
}

export type Geopoint = {
  _type: 'geopoint'
  lat?: number
  lng?: number
  alt?: number
}

export type Synthetic = {
  _id: string
  _type: 'synthetic'
  _createdAt: string
  _updatedAt: string
  _rev: string
  title?: string
  arrayOfObjects?: Array<
    {
      _key: string
    } & SyntheticObject
  >
  syntheticObject?: SyntheticObject
}

export type SyntheticObject = {
  _type: 'syntheticObject'
  name?: string
  image?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.imageAsset'
    }
    hotspot?: SanityImageHotspot
    crop?: SanityImageCrop
    _type: 'image'
  }
  file?: {
    asset?: {
      _ref: string
      _type: 'reference'
      _weak?: boolean
      [internalGroqTypeReferenceTo]?: 'sanity.fileAsset'
    }
    _type: 'file'
  }
  geopoint?: Geopoint
  number?: number
  string?: string
  boolean?: boolean
  slug?: Slug
  text?: string
  reference?: {
    _ref: string
    _type: 'reference'
    _weak?: boolean
    [internalGroqTypeReferenceTo]?: 'synthetic'
  }
  date?: string
  datetime?: string
  nestedObject?: SyntheticObject
  field0?: string
  field1?: string
  field2?: string
  field3?: string
  field4?: string
  field5?: string
  field6?: string
  field7?: string
  field8?: string
  field9?: string
  field10?: string
  field11?: string
  field12?: string
  field13?: string
  field14?: string
  field15?: string
  field16?: string
  field17?: string
  field18?: string
  field19?: string
}
// ...
export type SanityImageCrop = {
  _type: 'sanity.imageCrop'
  top?: number
  bottom?: number
  left?: number
  right?: number
}

export type SanityImageHotspot = {
  _type: 'sanity.imageHotspot'
  x?: number
  y?: number
  height?: number
  width?: number
}
// ...
export declare const internalGroqTypeReferenceTo: unique symbol